+++ Source/DotNetWorkQueue.Dashboard.Api.Tests/DashboardApiTests.cs
@@ -1,6 +1,8 @@
 using System;
 using DotNetWorkQueue.Dashboard.Api.Configuration;
 using FluentAssertions;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace DotNetWorkQueue.Dashboard.Api.Tests
@@ -12,7 +14,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void Connections_Empty_When_No_Registrations()
         {
             var options = new DashboardOptions();
-            using var api = new DashboardApi(options);
+            using var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
 
             api.Connections.Should().BeEmpty();
         }
@@ -21,7 +23,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void FindQueue_Returns_Null_For_Unknown_Id()
         {
             var options = new DashboardOptions();
-            using var api = new DashboardApi(options);
+            using var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
 
             var result = api.FindQueue(Guid.NewGuid());
 
@@ -32,7 +34,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void GetQueueContainer_Throws_For_Unknown_Id()
         {
             var options = new DashboardOptions();
-            using var api = new DashboardApi(options);
+            using var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
 
             var act = () => api.GetQueueContainer(Guid.NewGuid());
 
@@ -43,7 +45,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void Dispose_Can_Be_Called_Multiple_Times()
         {
             var options = new DashboardOptions();
-            var api = new DashboardApi(options);
+            var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
 
             api.Dispose();
             api.Dispose(); // should not throw
@@ -53,7 +55,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void FindQueue_Throws_After_Dispose()
         {
             var options = new DashboardOptions();
-            var api = new DashboardApi(options);
+            var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
             api.Dispose();
 
             var act = () => api.FindQueue(Guid.NewGuid());
@@ -65,7 +67,7 @@ namespace DotNetWorkQueue.Dashboard.Api.Tests
         public void GetQueueContainer_Throws_After_Dispose()
         {
             var options = new DashboardOptions();
-            var api = new DashboardApi(options);
+            var api = new DashboardApi(options, NullLogger<DashboardApi>.Instance);
             api.Dispose();
 
             var act = () => api.GetQueueContainer(Guid.NewGuid());